Entry-Test/backend/requirements.txt
Flask
scikit-learn
pandas
numpy
matplotlib
termcolor
openpyxl
requests
pymongo
mysql-connector-python
flask-mysql-connector
bcrypt
flask-jwt-extended